ci: switch release pipeline from semantic-release to Changesets
PRs add a .changeset/*.md file; on push to master, changesets/action
opens / updates a "Version Packages" PR. Merging that PR bumps
package.json, rewrites CHANGELOG.md, then runs scripts/release-publish.sh
to tag v$VERSION and create the GitHub Release. The existing binaries,
ftw-connect, ftw-subetha, docker, rpi-image, and discord jobs are
unchanged — they now gate on the changesets-action `published` output
and read the bumped version from package.json.

A new changeset-check workflow requires every PR to master to either
include a changeset or carry the `no-changeset` label.

The Hitchhiker codename header is preserved by porting the old
semantic-release plugin into scripts/apply-codename.cjs, invoked from
the publish step.
