--- .releaserc.json
@@ -1,84 +1,0 @@
-{
-  "branches": ["master"],
-  "plugins": [
-    [
-      "@semantic-release/commit-analyzer",
-      {
-        "preset": "conventionalcommits",
-        "releaseRules": [
-          { "type": "feat",     "release": "minor" },
-          { "type": "fix",      "release": "patch" },
-          { "type": "perf",     "release": "patch" },
-          { "type": "revert",   "release": "patch" },
-          { "type": "refactor", "release": "patch" },
-
-          { "type": "drivers",   "release": "minor" },
-          { "type": "driver",    "release": "minor" },
-          { "type": "ui",        "release": "minor" },
-          { "type": "control",   "release": "patch" },
-          { "type": "mpc",       "release": "patch" },
-          { "type": "telemetry", "release": "patch" },
-          { "type": "tsdb",      "release": "patch" },
-          { "type": "safety",    "release": "patch" },
-          { "type": "support",   "release": "minor" },
-
-          { "type": "docs",  "release": false },
-          { "type": "test",  "release": false },
-          { "type": "style", "release": false },
-          { "type": "chore", "release": false },
-          { "type": "build", "release": false },
-          { "type": "ci",    "release": false },
-
-          { "breaking": true, "release": "major" }
-        ],
-        "parserOpts": {
-          "noteKeywords": ["BREAKING CHANGE", "BREAKING-CHANGE", "BREAKING CHANGES"],
-          "headerPattern": "^(\\w*)(?:\\(([\\w$.\\-*/ ]*)\\))?!?: (.*)$",
-          "headerCorrespondence": ["type", "scope", "subject"]
-        }
-      }
-    ],
-    [
-      "@semantic-release/release-notes-generator",
-      {
-        "preset": "conventionalcommits",
-        "presetConfig": {
-          "types": [
-            { "type": "feat",      "section": "Features" },
-            { "type": "fix",       "section": "Bug Fixes" },
-            { "type": "perf",      "section": "Performance" },
-            { "type": "revert",    "section": "Reverts" },
-            { "type": "refactor",  "section": "Refactoring" },
-            { "type": "drivers",   "section": "Drivers" },
-            { "type": "driver",    "section": "Drivers" },
-            { "type": "ui",        "section": "UI" },
-            { "type": "control",   "section": "Control loop" },
-            { "type": "mpc",       "section": "MPC planner" },
-            { "type": "telemetry", "section": "Telemetry" },
-            { "type": "tsdb",      "section": "TSDB" },
-            { "type": "safety",    "section": "Safety" },
-            { "type": "support",   "section": "Device support" },
-            { "type": "docs",      "hidden": true },
-            { "type": "test",      "hidden": true },
-            { "type": "style",     "hidden": true },
-            { "type": "chore",     "hidden": true },
-            { "type": "build",     "hidden": true },
-            { "type": "ci",        "hidden": true }
-          ]
-        }
-      }
-    ],
-    "./scripts/release-codename.cjs",
-    [
-      "@semantic-release/github",
-      {
-        "successComment": false,
-        "failComment": false,
-        "failTitle": false,
-        "labels": false,
-        "releasedLabels": false,
-        "addReleases": false
-      }
-    ]
-  ]
-}
