+++ package.json
@@ -1,0 +1,14 @@
+{
+  "name": "forty-two-watts",
+  "version": "0.102.3",
+  "private": true,
+  "description": "Unified Home Energy Management System — version metadata only; the application itself is Go. Changesets reads `version` from this file to drive the release tag.",
+  "scripts": {
+    "changeset": "changeset",
+    "changeset:version": "changeset version",
+    "changeset:status": "changeset status --since=origin/master"
+  },
+  "devDependencies": {
+    "@changesets/cli": "^2.27.10"
+  }
+}
